refactor(ui): extract Mantine theme and style imports into theme module
