+++ eventum/ui/src/main.tsx
@@ -1,4 +1,3 @@
-import '@fontsource-variable/inter';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
